% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_synthesize.R
\name{ms_synthesize}
\alias{ms_synthesize}
\alias{ms_tts_url}
\title{Convert Text to Speech by using Speech Synthesis Markup Language (SSML)}
\usage{
ms_synthesize(
  script,
  region = "westus",
  api_key = NULL,
  token = NULL,
  gender = c("Female", "Male"),
  language = "en-US",
  voice = NULL,
  escape = FALSE
)

ms_tts_url(region = "westus")
}
\arguments{
\item{script}{A character vector of text to be converted to speech}

\item{region}{Subscription region for API key. For more info, see
\url{https://learn.microsoft.com/en-us/azure/cognitive-services/speech-service/regions}}

\item{api_key}{Microsoft Azure Cognitive Services API key}

\item{token}{An authentication token}

\item{gender}{Sex of the speaker}

\item{language}{Language to be spoken}

\item{voice}{Full voice name}

\item{escape}{Should non-standard characters be substituted?}
}
\value{
A response in binary format
}
\description{
Convert Text to Speech by using Speech Synthesis Markup Language (SSML)

Create Text To Speech Endpoint
}
