% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_voice.R
\name{ms_use_voice}
\alias{ms_use_voice}
\title{Check if provided voice is supported in region}
\usage{
ms_use_voice(voice, api_key = NULL, region = "westus")
}
\arguments{
\item{voice}{Full voice name}

\item{api_key}{Microsoft Azure Cognitive Services API key}

\item{region}{Subscription region for API key. For more info, see
\url{https://learn.microsoft.com/en-us/azure/cognitive-services/speech-service/regions}}
}
\value{
List of gender, language, and full voice name
}
\description{
Check if provided voice is supported in region
}
