% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{play_audio}
\alias{play_audio}
\title{Play audio in a browser}
\usage{
play_audio(audio = "output.wav", html = "player.html")
}
\arguments{
\item{audio}{The file location of the audio file.  Must be supported by HTML5}

\item{html}{The html file location that will be created host the audio}
}
\description{
This uses HTML5 audio tags to play audio in your browser. Borrowed from
\code{googleLanguageR::gl_talk_player()}.
}
\details{
For more info, see this \href{https://developer.mozilla.org/en-US/docs/Web/HTML/Element/audio}{Mozilla documentation}
detailing the \verb{<audio>} HTML element.
}
\examples{
\dontrun{
# Opens a browser with embedded audio
play_audio(audio = "output.wav")
}
}
