% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region}
\alias{region}
\title{Speech Service supported Region Identifier}
\format{
\subsection{\code{region}}{

A character vector
}
}
\source{
\url{https://learn.microsoft.com/en-us/azure/cognitive-services/Speech-Service/regions}
}
\usage{
region
}
\description{
This character vector contains region identifiers that support text to speech.
}
\keyword{datasets}
