% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_authorization.R
\name{ms_get_token}
\alias{ms_get_token}
\alias{ms_token_url}
\alias{ms_token_expired}
\title{Get Microsoft Text To Speech (TTS) or Cognitive Services Token from API Key}
\usage{
ms_get_token(api_key = NULL, region = "westus")

ms_token_url(region = conrad::region)

ms_token_expired(token = NULL)
}
\arguments{
\item{api_key}{Microsoft Azure Cognitive Services API key}

\item{region}{Subscription region for API key. For more info, see
\url{https://learn.microsoft.com/en-us/azure/cognitive-services/speech-service/regions}}

\item{token}{An authentication of class \code{token},
likely from \code{\link{ms_get_token}}}
}
\value{
A list of the request and token
}
\description{
Get Microsoft Text To Speech (TTS) or Cognitive Services Token from API Key

Create issueToken URL Endpoint to get access token

Check if token has expired
}
\examples{
if (ms_valid_key()) {
   token = ms_get_token()
}
}
