% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_region.R
\name{ms_region}
\alias{ms_region}
\alias{ms_set_region}
\title{Check if provided region is supported by Speech service}
\usage{
ms_region(region = mscstts2::region)

ms_set_region(region = mscstts2::region)
}
\arguments{
\item{region}{Subscription region for API key. For more info, see
\url{https://learn.microsoft.com/en-us/azure/cognitive-services/speech-service/regions}}
}
\value{
region
}
\description{
Check if provided region is supported by Speech service
}
