% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_synthesize.R
\name{ms_synthesize}
\alias{ms_synthesize}
\alias{ms_tts_url}
\title{Convert Text to Speech by using Speech Synthesis Markup Language (SSML)}
\usage{
ms_synthesize(
  script,
  region = "westus",
  api_key = NULL,
  token = NULL,
  gender = c("Female", "Male"),
  language = "en-US",
  voice = NULL,
  output_format = c("raw-16khz-16bit-mono-pcm", "raw-8khz-8bit-mono-mulaw",
    "riff-8khz-8bit-mono-alaw", "riff-8khz-8bit-mono-mulaw", "riff-16khz-16bit-mono-pcm",
    "audio-16khz-128kbitrate-mono-mp3", "audio-16khz-64kbitrate-mono-mp3",
    "audio-16khz-32kbitrate-mono-mp3", "raw-24khz-16bit-mono-pcm",
    "riff-24khz-16bit-mono-pcm", "audio-24khz-160kbitrate-mono-mp3",
    "audio-24khz-96kbitrate-mono-mp3", "audio-24khz-48kbitrate-mono-mp3"),
  escape = FALSE
)

ms_tts_url(region = "westus")
}
\arguments{
\item{script}{A character vector of text to be converted to speech}

\item{region}{Subscription region for API key. For more info, see
\url{https://learn.microsoft.com/en-us/azure/cognitive-services/speech-service/regions}}

\item{api_key}{Microsoft Azure Cognitive Services API key}

\item{token}{An authentication token}

\item{gender}{Sex of the speaker}

\item{language}{Language to be spoken}

\item{voice}{Full voice name}

\item{output_format}{Format of output}

\item{escape}{Should non-standard characters be substituted?}
}
\value{
A response in binary format
}
\description{
Convert Text to Speech by using Speech Synthesis Markup Language (SSML)

Create Text To Speech Endpoint
}
