% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_region.R, R/ms_synthesize.R
\name{ms_set_region}
\alias{ms_set_region}
\alias{ms_synthesize}
\alias{ms_synthesize_api_url}
\title{Convert Text to Speech by using Speech Synthesis Markup Language (SSML)}
\usage{
ms_set_region(region = mscstts2::region)

ms_synthesize(
  script,
  token = NULL,
  api_key = NULL,
  gender = c("Female", "Male"),
  language = "en-US",
  voice = NULL,
  output_format = c("raw-16khz-16bit-mono-pcm", "raw-8khz-8bit-mono-mulaw",
    "riff-8khz-8bit-mono-alaw", "riff-8khz-8bit-mono-mulaw", "riff-16khz-16bit-mono-pcm",
    "audio-16khz-128kbitrate-mono-mp3", "audio-16khz-64kbitrate-mono-mp3",
    "audio-16khz-32kbitrate-mono-mp3", "raw-24khz-16bit-mono-pcm",
    "riff-24khz-16bit-mono-pcm", "audio-24khz-160kbitrate-mono-mp3",
    "audio-24khz-96kbitrate-mono-mp3", "audio-24khz-48kbitrate-mono-mp3"),
  escape = FALSE,
  region = NULL,
  api = c("tts", "bing"),
  ...
)

ms_synthesize_api_url(api = c("tts", "bing"), region = NULL)
}
\arguments{
\item{api}{Chose API to authorize on (\code{tts} for text to speech or
\code{bing} for Bing text to speech API)}
}
\description{
Convert Text to Speech by using Speech Synthesis Markup Language (SSML)

Create URL Endpoint that allows you to convert text to speech
}
